리뷰, 댓글 완성

+++ src/components/products/ProductDetail분리전.vue
@@ -117,155 +117,8 @@
           </div>
         </div>
         <!-- 제품 정보 끝 -->
-
-        <!-- 제품 리뷰 시작 -->
-        <div class="border" style="margin-top:20px;">
-          <div style="margin-left:1em; margin-top:1em;">
-            <h4 style="font-weight:900;color:#3d3935;">제품 리뷰</h4>
-          </div>
-          <hr>
-          <!-- 로그인 되어 있는 경우(300자 까지 허용, 0/300으로 글자 수 카운트하기) -->
-          <div class="form-group" style="border:1px solid lightgray" v-if="isLogin">
-            <!-- rating start -->
-            <div class="text-center">
-              <span>제품 평가를 남겨주세요(최대 300자까지 가능합니다). </span>
-              <div class="text-center">
-                <span class="col-lg-2 font-weight-bold">리뷰제목</span><input type="text" v-model="review.title"
-                  class="form-control col-lg-12" ref="reviewTitle">
-              </div>
-              <label for="comment"><span class="font-weight-bold">글내용</span>(글자수 <span
-                  style="margin-left:0.5rem">{{ review.contents.length }}/300)</span></label>
-            </div>
-            <!-- <div class="container">
-              <div class="starrating risingstar d-flex justify-content-center flex-row-reverse" v-on:click="rating">
-                <input type="radio" id="star5" name="rating" value="5" /><label for="star5" title="5 star"></label>
-                <input type="radio" id="star4" name="rating" value="4" /><label for="star4" title="4 star"></label>
-                <input type="radio" id="star3" name="rating" value="3" /><label for="star3" title="3 star"></label>
-                <input type="radio" id="star2" name="rating" value="2" /><label for="star2" title="2 star"></label>
-                <input type="radio" id="star1" name="rating" value="1" /><label for="star1" title="1 star"></label>
-              </div>
-            </div> -->
-            <!-- rating end -->
-            <textarea class="form-control" rows="5" id="comment" v-model="review.contents" maxlength="300"
-              placeholder="최대 300까지 작성하실 수 있습니다." ref="reviewContents"></textarea>
-            <!-- 버튼은 오른쪽으로 배치(text-right) -->
-            <div style="higth:200px;" class="text-right">
-              <button type="button" class="btn btn-info btn-xl" @click="write">등록</button>
-            </div>
-          </div>
-          <!-- 로그인 되어 있지 않은 경우 -->
-          <div class="text-center col-lg-12" v-if="!isLogin">
-            <router-link :to="{ name: 'login' }" class="text-danger font-weight-bold font" style="font-size:1rem;">
-              회원만 제품 리뷰를 작성하실 수 있습니다.
-            </router-link>
-            <hr class="divider">
-          </div>
-          <!-- review list start -->
-          <!-- 별점 코드 :class rate적용하기(class가 2개인게 정상 맞다)  -->
-          <div class="container">
-            <h2 class="text-center" v-if="reviewList.length < 1"> 리뷰가 없습니다.</h2>
-
-            <!-- 리뷰는 최신순으로 정렬 -->
-            <div class="card" v-for="(reviews, idx) in reviewList" :key="reviews.id" ref="list">
-              <div class="card-body">
-                <div class="row">
-                  <div class="col-md-2">
-                    <img src="https://image.ibb.co/jw55Ex/def_face.jpg" class="img img-rounded img-fluid" />
-                    <p class="text-secondary text-center">{{ reviews.user_id }}</p>
-                  </div>
-                  <div class="col-md-10">
-                    <p>
-                      <h3 class="text-dark font-weight-bold">{{ reviews.title }}</h3>
-                      <!-- <span class="float-right"><i class="fa fa-star star rate"></i></span>
-                      <span class="float-right"><i class="fa fa-star star rate"></i></span>
-                      <span class="float-right"><i class="fa fa-star star rate"></i></span>
-                      <span class="float-right"><i class="fa fa-star star rate"></i></span>
-                      <span class="float-right"><i class="fa fa-star star rate"></i></span> -->
-
-                      <!-- </p> -->
-                      <span>{{ reviews.createdAt }}</span>
-                      <div class="clearfix"></div>
-                      <p class="text-dark font-weight-bolder">{{ reviews.contents }}</p>
-                      <p>
-                        <!-- 이 아래 코드는 대댓글에 적용 -->
-                        <!-- <router-link class="float-right btn btn-outline-primary ml-2" v-if="!isLogin" tag="button"
-                          :to="{ name: 'login' }" href="#" onclick="return false;"> <i class="fa fa-reply"></i> 댓글달기
-                        </router-link> -->
-                        <button class="float-right btn btn-outline-primary ml-2" @click="isVisible(idx)"><i
-                            class="fa fa-reply"></i> 댓글
-                        </button>
-                        <!-- <button class="float-right btn btn-outline-warning ml-2" @click="modifyReview(reviews.id, idx)" ><i
-                            class="fa fa-pencil-alt"></i> 수정
-                        </button> -->
-                        <button class="float-right btn btn-outline-danger ml-2" @click="deleteReview(reviews.id, idx)"><i
-                            class="fa fa-trash-alt"></i> 삭제
-                        </button>
-                        <!-- <button class="float-right btn btn-outline-primary ml-2" type="button" v-if="isLogin" tag="button" href="#" onclick="return false;" > <i class="fa fa-reply"></i> 댓글달기!</button> -->
-                        <!-- <button class="float-right btn btn-outline-primary ml-2" type="button" v-if="isLogin" href="#"
-                          onclick="return false;" @click="isVisible"><i class="fa fa-reply"></i> 댓글보기</button> -->
-                      </p>
-                  </div>
-                </div>
-                <!-- 댓글 시작 -->
-                <!-- ref="test" :class="{ visible: true }"를 어디에 넣어야 하지? -->
-                <!-- <div class="card" ref="test" :class="{ visible: true }" v-if="list.temp"> -->
-                <div class="card" ref="test" v-if="review.temp">
-                  <!-- 댓글 입력폼 시작 -->
-                  <!-- 댓글작성도 로그인 한 사용자만 볼 수 있어야 한다. -->
-                  <!-- 대댓글 작성한다고 하면 안 보이게 할까? 입력창이 2개니까 헷갈린다 -->
-                  <div class="form-group" style="border:1px solid lightgray" v-if="isLogin" >
-                    <div class="text-center">
-                      <span>댓글 작성(최대 300자까지 가능합니다). </span>
-                    </div>
-                    <textarea class="form-control" rows="5" id="comment" maxlength="300"
-                      placeholder="최대 300까지 작성하실 수 있습니다." ref="comment"></textarea>
-                    <!-- 버튼은 오른쪽으로 배치(text-right) -->
-                    <div style="higth:200px;" class="text-right">
-                      <!-- 리뷰의 id를 전달해 줘야 함 -->
-                      <button type="button" class="btn btn-secondary btn-xl" @click="reply(reviews.id, idx)">댓글등록</button>
-                    </div>
-                  </div>
-                  <!-- 댓글 입력폼 끝 -->
-                  <!-- 댓글 보기 시작 -->
-
-                  <!-- <h4 class="text-center font-bolder text-body" v-show="typeof item.comments.length === 'undefined' || item.comments.length < 1 ">댓글이 없습니다</h4> -->
-                  <h4 class="text-center font-bolder text-body" v-if="reviews.comments.length < 1">댓글이 없습니다</h4>
-                  <!-- 새로 작성한 글은 코멘트 객체가 없으므로 여기서 에러 발생 -->
-                  <div class="card-body border" v-for="(one, idx) in reviews.comments" :key="idx">
-                    <div class="row" >
-                      <div class="col-md-12">
-                        <p>
-                          <!-- <span class="float-right"><i class="fa fa-star star rate"></i></span>
-                      <span class="float-right"><i class="fa fa-star star rate"></i></span>
-                      <span class="float-right"><i class="fa fa-star star rate"></i></span>
-                      <span class="float-right"><i class="fa fa-star star rate"></i></span>
-                      <span class="float-right"><i class="fa fa-star star rate"></i></span> -->
-
-                          <!-- </p> -->
-                          <span>{{ one.createdAt }}</span>
-                          <div class="clearfix"></div>
-                          <!-- <strong class="text-dark font-bolder">@수신자</strong> -->
-                          <p class="text-dark font-weight-bolder">{{ one.contents }}</p>
-                          <p>
-                           <!-- 자기가 쓴 글일 경우 수정, 삭제가 보이게 -->
-                            <button class="float-right btn btn-outline-warning ml-2" v-if="id === one.userId && isLogin"><i
-                                class="fas fa-pencil-alt"></i> 수정
-                            </button>
-                            <button class="float-right btn btn-outline-danger ml-2" v-if="id === one.userId && isLogin" 
-                            ><i class="fas fa-trash-alt"></i> 삭제
-                            </button>
-                          </p>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <!-- 댓글 끝 -->
-              </div>
-            </div>
-          </div>
-          <!-- review list end -->
-        </div>
-        <!-- 제품 리뷰 끝 -->
+        <review/>
+      
       </div>
     </div>
   </div>
@@ -273,6 +126,9 @@
 
 <script>
   import { mapState, mapGetters } from 'vuex'
+  import Review from './Review.vue'
+  // eslint-disable-next-line no-unused-vars
+  import store from '@/store'
   export default {
     name: 'ProductDetail',
     data() {
@@ -309,18 +165,38 @@
       //   var ele = event.target.value;
       //   console.log(ele);
       // },
-      isVisible(idx) {
-        // 댓글보기 활성화/비활성화 여부
+      isVisible(idx, id) {
+        // 댓글보기 활성화/비활성화 여부(id는 리뷰의id)
         // 그리고 댓글 불러오기
         console.log('몇번째요소?')
         console.log(idx);
-        // !(this.$refs.test[idx].style.display)
-        //   ? this.$refs.test[idx].style.display = 'block'
-        //   : this.$refs.test[idx].style.display = ''
-        !(this.$refs.test[idx].style.display)
-          ? this.$refs.test[idx].temp = true
-          : this.$refs.test[idx].style.display = false
-        
+        console.log(this.$refs.list[idx]);
+        // 아래 코드는 유효x
+        // !(this.$refs.list[idx].style.display)
+        //   ? this.$refs.list[idx].style.display = ''
+        //   : this.$refs.list[idx].style.display = ''
+        (async () => {
+          try {
+            const response = await this.$axios.get(`${ this.url }/reviews/comments?review_id=${ id }`);
+            console.log('가져온 댓글', response.data);
+
+            console.log(this.reviewList[idx])
+            this.reviewList[idx].comments = response.data;
+            response.data.length > 0
+              ? this.reviewList[idx].commentLength = ''
+              : this.reviewList[idx].commentLength = '댓글이 없습니다.'
+            this.$forceUpdate();
+           
+          } catch(err) {
+            console.log(err);
+          }
+        })();
+        this.$forceUpdate();
+        // 클릭에 의해 댓글창이 보일지 말지 결정
+        this.reviewList[idx].isShow 
+          ? this.reviewList[idx].isShow = false
+          : this.reviewList[idx].isShow = true
+
       },
       write() {
         (async () => {
@@ -345,7 +221,9 @@
               // 방금 생성된 상품평에는 코멘트가 없어서 v-for로 뿌릴 때 에러 발생
               // 에러를 막기 위해 빈 코맨드 배열 추가
               response.data.result.comments = [];
-
+              this.review.title = '';
+              this.review.contents = '';
+              
               // splice: 배열의 기존 요소를 삭제 또는 교체하거나
               //  새 요소를 추가하여 배열의 내용을 변경합니다.
               // array.splice(start, deleteCount, item )
@@ -357,19 +235,22 @@
 
               console.log('리스트 바뀌나?')
               console.log(this.reviewList);
-              
-              // this.reviewList = result.data.reviewInfo;
+              // 반드시 강제로 재 렌더링을 해줘야 vue에서 제대로 인식됨
+              // 안 그러면 댓글 달 때도 인식X(예전 리스트를 인식함)
+              this.$forceUpdate();
+
             } else {
               alert(message);
             }
           } catch (err) {
             console.log(err);
           }
-        }
-        )();
+        })();
       },
       deleteReview(id, idx){
-        (async () => {
+        const isDelete = confirm('삭제하시겠습니까?');
+        if(isDelete) {
+          (async () => {
             this.reviewList.splice(idx, 1);
             try {
               const response = this.$axios.delete(`${ this.url }/reviews/delete?review_id=${ id }`)
@@ -377,7 +258,10 @@
             } catch(err) {
               console.log(err);
             }
-        })();
+          })();
+        } else {
+          return false;
+        }
       },
       // modifyReview(id, idx){
       //   (async () => {
@@ -393,25 +277,48 @@
       // },
       reply(reviewId, idx){
         (async () => {
-          try {
+         
+            console.log('리뷰id확인', reviewId);
+            console.log('작성자 id확인', this.comment.userId)
+            console.log('인덱스확인', idx);
+            console.log('댓글요소', this.$refs.comment[idx]);
+            console.log('댓글내용확인', this.$refs.comment[idx].value);
+            console.log('리뷰대상', this.reviewList[idx]);
+            // console.log('dom에 접근한다', this.$refs.list[idx]);
+            // console.log('자식에 접근한다', this.$refs.list[idx].childNodes);
+            // console.log('자식에 접근한다', this.$refs.list[idx].childNodes.childNodes);
+            // console.log('댓글내용확인', this.$refs.comment[idx].value);
+
+            // v-for로 반복을 돌렸으므로 v-model로는 처리불가
+            // data옵션에 정의하면 모든 v-for요소들이 상태를 공유하게 된다.
             this.comment.reviewId = reviewId;
             this.comment.userId = this.id;
             this.comment.contents = this.$refs.comment[idx].value;
-            console.log('리뷰id확인');
-            console.log(reviewId);
-            console.log('작성자 id확인')
-            console.log(this.comment.writer)
-            console.log('댓글내용확인');
-            console.log(this.$refs.comment[idx].value);
            
-            const result = await this.$axios.post(`${ this.url }/reviews/comments`, this.comment);
-            console.log(result);
-            const { code, message } = result.data;
+          if(!this.comment.contents) {
+             alert('댓글을 입력해주세요');
+             return false;
+          }
+          try {
+            const response = await this.$axios.post(`${ this.url }/reviews/comments`, this.comment);
+            console.log(response);
+            const { code, message } = response.data;
 
             if(code === 201) {
               // 받아온 데이터를 기존 댓글 리스트에 추가해야한다.
-              // 리스트가 갱신되었음에도 화면에는 추가되지 않는다. 왜??
-              this.reviewList[idx].comments.push(result.data.result)
+              // user_id가 없음. 서버에서 받아와야 함.
+              console.log('새로 작성한 코멘트 받아옴', response.data.result)
+              // 댓글 단 사람에 현재 user_id추가해주기
+              console.log('store의serid확인', this.userId);
+              response.data.result.user_id = this.userId;
+              console.log('user_id추가 확인', response.data.result)
+              
+              this.reviewList[idx].comments.push(response.data.result);
+              this.comment.contents = '';
+              this.$refs.comment[idx].value = '';
+              // reviewId도 지워줘야 함. 지우지 않을시 에러 발생. 왜?
+              this.comment.reviewId = '';
+              this.$forceUpdate();
             } else {
               alert(message);
             }
@@ -427,14 +334,17 @@
       // 경로를 직접 입력하여 들어간 경우 생명주기가 작동하지 않는걸로 보인다.
       // 동일한 컴포넌트 인스턴스가 재사용 되기 때문으로 보임.
       (async () => {
+        // 먼저 현재 로그인 유무, 유효성을 확인한다.
         try {
-          const result = await this.$axios.get(`${this.url}/products/${this.$route.params.id}`)
+          const result = await this.$axios.get(`${ this.url }/products/${ this.$route.params.id }`)
           console.log(result);
           this.product = result.data.productInfo;
           this.reviewList = result.data.reviewInfo;
           // 댓글 더보기에 사용할 요소 추가
           this.reviewList.forEach(list => {
-            list.temp = false;
+            list.isShow = false
+            list.comments = [];
+            list.commentLength = ''; // 댓글이 없습니다에 사용될 변수
           })
           console.log('확인해보자!!', this.reviewList);
 
@@ -447,8 +357,8 @@
 
           console.log(this.product)
           console.log('detail에서 usreid확인**');
-          console.log(this.review.userId);
-
+          console.log(this.review.userId, 'store는', this.id);
+          console.log('productDetail에서 user_id확인', this.user_id);
           console.log('product params확인')
           console.log(this.$route.params);
         } catch (err) {
@@ -486,13 +396,11 @@
 
     },
     beforeRouteEnter(to, from, next) {
+      // 글 작성을 위해 로그인 여부, id를 확인한다.
+      store.dispatch('getUserId');
       next()
     },
     watch: {
-      'reviewList.length': function(newVal){
-        console.log('reviewListLength의 길이 변경됨')
-        return newVal;
-      }
     },
     filters: {
       parent_category(val) {
@@ -503,8 +411,11 @@
         }
       },
     },
+    components: {
+      review: Review
+    },
     computed: {
-      ...mapState(['isLogin', 'id', 'url']),
+      ...mapState(['isLogin', 'id', 'userId', 'url']),
       ...mapGetters(['productImagePath', 'pictogramImagePath']),
       child_category() {
         // filter의 this는 undefined
@@ -613,8 +524,4 @@
     color: orange;
   }
 
-  /* 댓글 창 보일지 여부 */
-  .visible {
-    display: none;
-  }
 </style>
